requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==4.2.26
django-cors-headers==4.9.0
djangorestframework==3.16.1
fonttools==4.60.1
idna==3.11
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyparsing==3.2.5
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.17.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
QtAwesome==1.4.0
QtPy==2.4.3
reportlab==4.4.4
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
